pyproject.toml
[tool.poetry]
name = "bomb-bot"
version = "0.1.0"
description = ""
authors = ["henryleecode23 <99625462+hlc23@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-cord = {extras = ["speed"], version = "^2.4.1"}
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
